add models module
