+++ src/pydisagg/disaggregate.py
@@ -5,8 +5,7 @@ from typing import Literal
 import numpy as np
 import pandas as pd
 
-from pydisagg.DisaggModel import DisaggModel
-from pydisagg.models import LogOddsModel
+from pydisagg.models import DisaggModel, LogOddsModel
 from pydisagg.typing import DataFrame, NDArray
