+++ src/pydisagg/models/__init__.py
@@ -1,0 +1,6 @@
+from pydisagg.models.disagg_model import DisaggModel
+from pydisagg.models.lmo_model import LMOModel
+from pydisagg.models.log_odds_model import LogOddsModel
+from pydisagg.models.rate_multiplicative_model import RateMultiplicativeModel
+
+__all__ = ["DisaggModel", "LMOModel", "LogOddsModel", "RateMultiplicativeModel"]
