+++ src/pydisagg/models/disagg_model.py
