+++ src/pydisagg/models/rate_multiplicative_model.py
@@ -1,10 +1,8 @@
-"""Module containing specific splitting models with transformations built in"""
-
 import numpy as np
 from numpy.typing import NDArray
 
-from pydisagg.DisaggModel import DisaggModel
-from pydisagg.transformations import Log, LogModifiedOdds, LogOdds
+from pydisagg.models.disagg_model import DisaggModel
+from pydisagg.transformations import Log
 
 
 class RateMultiplicativeModel(DisaggModel):
@@ -34,17 +32,3 @@ class RateMultiplicativeModel(DisaggModel):
             observed_total / np.sum(bucket_populations * rate_pattern)
         )
         return beta_val
-
-
-class LMOModel(DisaggModel):
-    """DisaggModel using the log-modified odds transformation with the exponent m."""
-
-    def __init__(self, m: float) -> None:
-        super().__init__(transformation=LogModifiedOdds(m))
-
-
-class LogOddsModel(DisaggModel):
-    """Produces an DisaggModel assuming multiplicativity in the odds"""
-
-    def __init__(self) -> None:
-        super().__init__(transformation=LogOdds())
